project/requirements.txt
Flask==3.1.0
pymongo==4.11.1
pytesseract==0.3.13
Pillow==11.1.0
opencv-python-headless==4.10.0.84
pdf2image==1.17.0
pypdf==5.4.0
python-dateutil==2.9.0.post0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==23.0.0
